ci: fix clustermesh worfklows on stable branches
The clustermesh workflows currently fail on stable branches as they
expect `.github/kind-config.yaml.tmpl` to be present, but this file is
only present on `main`.

It seems the template file is the only reason we need to checkout the
repository, as we do not use any other file from the repository in
subsequent steps, so we fix this by removing the checkout of the current
branch since we already checkout the `main` branch in a prior step.

Thus, the template file on `main` is used for all worfkows.

Signed-off-by: Nicolas Busseneau <nicolas@isovalent.com>